switched to single precision floating point ('float' instead of 'double')

+++ annlib/activation_function.h
@@ -8,20 +8,20 @@ namespace annlib
 	class activation_function
 	{
 	public:
-		const std::function<double(double)> f;
-		const std::function<double(double)> df;
+		const std::function<float(float)> f;
+		const std::function<float(float)> df;
 
-		activation_function(std::function<double(double)> f, std::function<double(double)> df);
+		activation_function(std::function<float(float)> f, std::function<float(float)> df);
 
-		double apply(double d) const;
+		float apply(float d) const;
 
-		double apply_derivative(double d) const;
+		float apply_derivative(float d) const;
 	};
 
 	class logistic_activation_function : public activation_function
 	{
 	public:
-		explicit logistic_activation_function(double T);
+		explicit logistic_activation_function(float T);
 	};
 }
